requirements.txt
psycopg2==2.8.4
Django==2.2.6
pytz==2019.3
djangorestframework==3.10.3
Pillow==6.2.1
celery==4.3.0